+++ src/main/php/com/skplanet/syruppay/token/claims/elements/RegistrationRestrictions.php
@@ -21,46 +21,23 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
+
 namespace com\skplanet\syruppay\token\claims\elements;
+
 
 use com\skplanet\syruppay\token\PropertyMapper;
 
-class Accept extends PropertyMapper implements Element
+class RegistrationRestrictions extends PropertyMapper
 {
-    protected $type;
-    protected $conditions = array();
+    protected $matchedUser;
 
-    public function getType()
+    public function setMatchedUser($matchedUser)
     {
-        return $this->type;
+        $this->matchedUser = $matchedUser;
     }
 
-    public function setType($type)
+    public function getMatchedUser()
     {
-        $this->type = $type;
-        return $this;
-    }
-
-    public function getConditions()
-    {
-        return $this->conditions;
-    }
-
-    public function addConditions($cardCode, $minPaymentAmt)
-    {
-        $this->conditions[] = array('cardCode' => $cardCode, 'minPaymentAmt' => $minPaymentAmt);
-        return $this;
-    }
-
-    function validRequired()
-    {
-        if (!isset($type)) {
-            throw new \InvalidArgumentException("Accept object couldn't be with null fields.");
-        }
-
-        if (!isset($this->conditions) || empty($this->conditions)) {
-            throw new \InvalidArgumentException("Conditions of Accept object couldn't be empty. you should contain with conditions of Accept object.");
-        }
+        return $this->matchedUser;
     }
 }
-
